fix logo home button to use react router navigate
