+++ src/shared/Routes.js
@@ -1,7 +1,8 @@
 import React, { useEffect, useRef, useState } from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { useSelector } from "react-redux";
 
-import Layout from "./Layout";
+import Header from "./Header";
 import Home from "../pages/Home";
 import Login from "../pages/Login";
 import Detail from "../pages/Detail";
@@ -12,20 +13,22 @@ const Router = () => {
   const ref = useRef();
   useEffect(() => {
     if (!ref?.current) return;
+    if (window.innerHeight === 0) return;
     setBodyHeight(window.innerHeight - ref.current.clientHeight);
-  }, [ref]);
+  }, [window.innerHeight, ref]);
+
+  const isLogin = useSelector((store) => store.user.isLogin);
 
   return (
-    <Layout ref={ref}>
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Home minHeight={bodyHeight} />} />
-          <Route path="/login" element={<Login />} />
-          <Route path="/post/:id" element={<Detail minHeight={bodyHeight} />} />
-          <Route path="/post" element={<PostCreate />} />
-        </Routes>
-      </BrowserRouter>
-    </Layout>
+    <BrowserRouter>
+      <Header ref={ref} show={isLogin} />
+      <Routes>
+        <Route path="/" element={<Home minHeight={bodyHeight} />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/post/:id" element={<Detail minHeight={bodyHeight} />} />
+        <Route path="/post" element={<PostCreate />} />
+      </Routes>
+    </BrowserRouter>
   );
 };
